+++ tests/ie/testing-lib.tsx
@@ -1,0 +1,2 @@
+import "../utils/ie-polyfills";
+import "../testing-lib";