+++ tests/utils/delay.ts
@@ -1,0 +1,2 @@
+export default (time: number) =>
+  new Promise((resolve) => setTimeout(resolve, time));